9009: Move getConnection.ts from web-api/src to web-api/src/persistence/postgres and update all imports
