+++ web-api/src/persistence/postgres/getConnection3.test.ts
@@ -27,7 +27,7 @@ jest.mock('pg', () => {
     Pool,
   };
 });
-import { getConnection } from '@web-api/getConnection';
+import { getConnection } from '@web-api/persistence/postgres/getConnection';
 import { environment } from '@web-api/environment';
 import { sleep } from '@shared/tools/helpers';
 import { formatNow as formatNowMock } from '@shared/business/utilities/DateHandler';
